Created FeedItemMapper class in order to clean up feed image mapping from 'RemoteFeedLoader' load function.
