+++ FeedAPIChallenge/Feed API/RemoteFeedLoader.swift
@@ -22,29 +22,14 @@ public final class RemoteFeedLoader: FeedLoader {
         client.get(from: url) { result in
             switch result {
             case let .success((data, response)):
-                guard response.statusCode == 200, let root = try? JSONDecoder().decode(Root.self, from: data) else {
+                guard response.statusCode == 200, let items = FeedItemMapper.mapJSONToFeedImage(from: data) else {
                     completion(.failure(Error.invalidData))
                     return
                 }
-                completion(.success(root.items.map { $0.item }))
+                completion(.success(items))
             case .failure(_):
                 completion(.failure(Error.connectivity))
             }
         }
-    }
-}
-
-struct Root: Codable {
-    let items: [Item]
-}
-
-struct Item: Codable {
-    let image_id: UUID
-    let image_desc: String?
-    let image_loc: String?
-    let image_url: URL
-    
-    var item: FeedImage {
-        return FeedImage(id: image_id, description: image_desc, location: image_loc, url: image_url)
     }
 }
